chore: extract methods for metrics/inputs to integrations files
