+++ e2e/_suites/fleet/fleet.go
@@ -6,16 +6,9 @@ package main
 
 import (
 	"context"
-	"encoding/json"
-	"fmt"
-	"io/ioutil"
-	"os"
-	"path/filepath"
 	"time"
 
-	"github.com/Jeffail/gabs/v2"
 	"github.com/docker/go-connections/nat"
-	"github.com/google/uuid"
 
 	"github.com/cenkalti/backoff/v4"
 	"github.com/elastic/e2e-testing/internal/common"
@@ -232,118 +225,4 @@ func (fts *FleetTestSuite) theHostIsRestarted() error {
 
 	log.Debug("The elastic-agent service has been restarted")
 	return nil
-}
-
-func inputs(integration string) []kibana.Input {
-	switch integration {
-	case "apm":
-		return []kibana.Input{
-			{
-				Type:    "apm",
-				Enabled: true,
-				Streams: []kibana.Stream{},
-				Vars: map[string]kibana.Var{
-					"apm-server": {
-						Value: "host",
-						Type:  "localhost:8200",
-					},
-				},
-			},
-		}
-	case "linux":
-		return []kibana.Input{
-			{
-				Type:    "linux/metrics",
-				Enabled: true,
-				Streams: []kibana.Stream{
-					{
-						ID:      "linux/metrics-linux.memory-" + uuid.New().String(),
-						Enabled: true,
-						DS: kibana.DataStream{
-							Dataset: "linux.memory",
-							Type:    "metrics",
-						},
-						Vars: map[string]kibana.Var{
-							"period": {
-								Value: "1s",
-								Type:  "string",
-							},
-						},
-					},
-				},
-			},
-		}
-	}
-	return []kibana.Input{}
-}
-
-func metricsInputs(integration string, set string, file string, metrics string) []kibana.Input {
-	metricsFile := filepath.Join(testResourcesDir, file)
-	jsonData, err := readJSONFile(metricsFile)
-	if err != nil {
-		log.Warnf("An error happened while reading metrics file, returning an empty array of inputs: %v", err)
-		return []kibana.Input{}
-	}
-
-	data := parseJSONMetrics(jsonData, integration, set, metrics)
-	return []kibana.Input{
-		{
-			Type:    integration,
-			Enabled: true,
-			Streams: data,
-		},
-	}
-
-	return []kibana.Input{}
-}
-
-func readJSONFile(file string) (*gabs.Container, error) {
-	jsonFile, err := os.Open(file)
-	if err != nil {
-		fmt.Println(err)
-	}
-	log.WithFields(log.Fields{
-		"file": file,
-	}).Info("Successfully Opened " + file)
-
-	defer jsonFile.Close()
-	data, err := ioutil.ReadAll(jsonFile)
-	if err != nil {
-		return nil, err
-	}
-
-	jsonParsed, err := gabs.ParseJSON(data)
-	if err != nil {
-		return nil, err
-	}
-
-	return jsonParsed.S("inputs"), nil
-}
-
-func parseJSONMetrics(data *gabs.Container, integration string, set string, metrics string) []kibana.Stream {
-	for i, item := range data.Children() {
-		if item.Path("type").Data().(string) == integration {
-			for idx, stream := range item.S("streams").Children() {
-				dataSet, _ := stream.Path("data_stream.dataset").Data().(string)
-				if dataSet == metrics+"."+set {
-					data.SetP(
-						integration+"-"+metrics+"."+set+"-"+uuid.New().String(),
-						fmt.Sprintf("inputs.%d.streams.%d.id", i, idx),
-					)
-					data.SetP(
-						true,
-						fmt.Sprintf("inputs.%d.streams.%d.enabled", i, idx),
-					)
-
-					var dataStreamOut []kibana.Stream
-					if err := json.Unmarshal(data.Path(fmt.Sprintf("inputs.%d.streams", i)).Bytes(), &dataStreamOut); err != nil {
-						return []kibana.Stream{}
-					}
-
-					return dataStreamOut
-				}
-			}
-		}
-	}
-	return []kibana.Stream{}
 }
